Add non-diatonic support for nashville numbers
Cheating a bit here, and using default borrowed chords rules. More
properly, the use of a # or b version should be determined by the
relative direction of harmonic motion (# for upwards, b for downwards)

But that's complex, as it requires a context-aware parser. And using
common borrowed defaults gets us 95% of the way there, and still
functional, if not 100% conventionally correct.